+++ ChangeLog.md
@@ -2,6 +2,7 @@
 * Add Offending Element tags to C-Move response output
 * Add exception handling for C-Store requests with unparsable datasets
 * Handle ObjectDisposedExceptions in network operations
+* Fix transcoding between compressed transfer syntaxes
 
 #### v1.0.21 (09/14/2012)
 * Add connection close event and socket error handlers to DicomService